refactor: :truck: rename module from `exclude` to `exclusion`
