+++ src/check_datapackage/__init__.py
@@ -4,7 +4,7 @@ from .check import check
 from .config import Config
 from .custom_check import CustomCheck
 from .examples import example_package_properties, example_resource_properties
-from .exclude import Exclusion
+from .exclusion import Exclusion
 from .issue import Issue
 from .read_json import read_json
