+++ src/check_datapackage/config.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 from typing import Literal
 
 from check_datapackage.custom_check import CustomCheck
-from check_datapackage.exclude import Exclusion
+from check_datapackage.exclusion import Exclusion
 
 
 @dataclass
