+++ src/check_datapackage/exclusion.py
